Added ailia implementation
