+++ audio_processing/crnn_audio_classification/crnn_audio_classification_util.py
@@ -1,49 +1,3 @@
-import time
-import sys
-import argparse
-
-import numpy as np
-
-import ailia  # noqa: E402
-
-import torch
-from torchvision import transforms
-
-import onnxruntime
-
-# import original modules
-sys.path.append('../../util')
-from model_utils import check_and_download_models  # noqa: E402
-
-# ======================
-# Arguemnt Parser Config
-# ======================
-
-parser = argparse.ArgumentParser(
-    description='CRNN Audio Classification.'
-)
-parser.add_argument(
-    '-b', '--benchmark',
-    action='store_true',
-    help='Running the inference on the same input 5 times ' +
-         'to measure execution performance. (Cannot be used in video mode)'
-)
-
-args = parser.parse_args()
-
-
-# ======================
-# PARAMETERS
-# ======================
-
-WEIGHT_PATH = "crnn.onnx"
-MODEL_PATH = "crnn.onnx.prototxt"
-REMOTE_PATH = "https://storage.googleapis.com/ailia-models/crnn/"
-
-# ======================
-# Preprocess
-# ======================
-
 import numpy as np
 import torch
 import torch.nn as nn
@@ -51,6 +5,8 @@ import torch.nn as nn
 from torchaudio.transforms import Spectrogram, MelSpectrogram , ComplexNorm
 from torchaudio.transforms import TimeStretch, AmplitudeToDB 
 from torch.distributions import Uniform
+
+from torchvision import transforms
 
 def _num_stft_bins(lengths, fft_length, hop_length, pad):
     return (lengths + 2 * pad - fft_length + hop_length) // hop_length
@@ -192,113 +148,3 @@ class ToTensorAudio(object):
 
     def __repr__(self):
         return self.__class__.__name__ + '()'
-
-import soundfile as sf
-
-def load_audio(path):
-    return sf.read(path)
-
-def preprocess(batch):
-    spec = MelspectrogramStretch(hop_length=None, 
-                            num_mels=128, 
-                            fft_length=2048, 
-                            norm='whiten', 
-                            stretch_param=[0.4, 0.4])
-                            
-    # x-> (batch, time, channel)
-    x, lengths, _ = batch # unpacking seqs, lengths and srs
-    # x-> (batch, channel, time)
-    xt = x.float().transpose(1,2)
-    # xt -> (batch, channel, freq, time)
-    print(xt.shape)
-    xt, lengths = spec(xt, lengths)                
-    print(xt.shape)
-
-    #for key in self.net:
-    #    print(key)
-    
-    return xt, lengths
-
-
-def postprocess(out_raw):
-    classes = ['air_conditioner', 'car_horn', 'children_playing', 'dog_bark', 'drilling',
- 'engine_idling', 'gun_shot', 'jackhammer', 'siren', 'street_music']
-    out = torch.exp(out_raw)
-    max_ind = out.argmax().item()      
-    print(max_ind)  
-    print(out.shape)
-    print(len(classes))
-    return classes[max_ind], out[:,max_ind].item()
-
-# ======================
-# Main function
-# ======================
-
-def crnn(path):
-    # normal inference
-    data = load_audio(path)
-    tsf = AudioTransforms()
-    sig_t, sr, _ = tsf.apply(data, None)
-
-    length = torch.tensor(sig_t.size(0))
-    sr = torch.tensor(sr)
-    data = [d.unsqueeze(0).to("cpu") for d in [sig_t, length, sr]]
-    
-    #label, conf = self.model.predict( data )
-
-    xt, lengths = preprocess(data) 
-
-    # inference
-    session = onnxruntime.InferenceSession("crnn.onnx")
-    xt = xt.to('cpu').detach().numpy().copy()
-    lengths = lengths.to('cpu').detach().numpy().copy()
-    results = session.run(["conf"],{ "data": xt, "lengths": lengths})
-
-    x = torch.from_numpy(results[0].astype(np.float32)).clone()
-
-    label, conf = postprocess(x)
-
-    return label, conf
-
-def main():
-    path = "dog.wav" #dog_bark 0.5050086379051208
-    label, conf = crnn(path)
-    print(label)
-    print(conf)
-
-
-def main_ailia():
-    # model files check and download
-    #check_and_download_models(WEIGHT_PATH, MODEL_PATH, REMOTE_PATH)
-
-    # net initialize
-    #env_id = ailia.get_gpu_environment_id()
-    #print(f'env_id: {env_id}')
-    #net = ailia.Net(MODEL_PATH, WEIGHT_PATH, env_id=env_id)
-
-    # inference
-    print('Start inference...')
-    if args.benchmark:
-        print('BENCHMARK mode')
-        for c in range(5):
-            start = int(round(time.time() * 1000))
-            preds_ailia = net.predict(input_data)
-            end = int(round(time.time() * 1000))
-            print("\tailia processing time {} ms".format(end-start))
-    else:
-        preds_ailia = net.predict(input_data)
-
-    # masked word prediction
-    predicted_indices = np.argsort(
-        preds_ailia[0][0][masked_index]
-    )[-NUM_PREDICT:][::-1]
-
-    predicted_tokens = tokenizer.convert_ids_to_tokens(predicted_indices)
-
-    print('Input sentence: ' + SENTENCE)
-    print(f'predicted top {NUM_PREDICT} words: {predicted_tokens}')
-    print('Script finished successfully.')
-
-
-if __name__ == "__main__":
-    main()
